requirements.txt
appdirs==1.4.4+computecanada
audioread==3.0.0+computecanada
certifi==2022.12.7+computecanada
cffi==1.15.1+computecanada
charset_normalizer==3.1.0+computecanada
click==8.1.3+computecanada
docker-pycreds==0.4.0+computecanada
filelock==3.12.0+computecanada
fsspec==2023.4.0+computecanada
gitdb==4.0.10+computecanada
GitPython==3.1.31+computecanada
huggingface-hub==0.14.1
HyperPyYAML==1.2.0
idna==3.4+computecanada
joblib==1.2.0+computecanada
lazy_loader==0.2+computecanada
librosa==0.10.0.post2+computecanada
llvmlite==0.39.1+computecanada
msgpack==1.0.4+computecanada
numba==0.56.3+computecanada
numpy==1.23.2+computecanada
packaging==23.1+computecanada
pathtools==0.1.2+computecanada
pooch==1.6.0+computecanada
protobuf==4.22.3
pycparser==2.21+computecanada
PyYAML==6.0+computecanada
regex==2022.10.31+computecanada
requests==2.30.0
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7+computecanada
scikit_learn==1.2.1+computecanada
scipy==1.10.1+computecanada
sentencepiece==0.1.97+computecanada
sentry-sdk==1.21.1
setproctitle==1.3.2+computecanada
smmap==5.0.0+computecanada
soundfile==0.12.1+computecanada
soxr==0.3.5+computecanada
speechbrain==0.5.14
threadpoolctl==3.1.0+computecanada
tokenizers==0.13.2+computecanada
torch==1.13.1+computecanada
torchaudio==0.13.1+computecanada
tqdm==4.65.0+computecanada
transformers==4.28.1+computecanada
typing_extensions==4.5.0+computecanada
urllib3==2.0.2
wandb==0.15.1
